Move store_names() to nif_import
Just to get it out of armature, but will be deleted soon
