+++ io_scene_nif/nif_import.py
@@ -293,7 +293,7 @@ class NifImport(NifCommon):
 
         # store original names for re-export
         if self.dict_names:
-            self.armaturehelper.store_names()
+            self.store_names()
         
         
         # now all havok objects are imported, so we are
@@ -1207,3 +1207,22 @@ class NifImport(NifCommon):
         return [ child for child in niBlock.children
                  if (isinstance(child, NifFormat.NiTriBasedGeom)
                      and child.name.find(node_name) != -1) ]
+
+
+    def store_names(self):
+        """Stores the original, long object names so that they can be
+        re-exported. In order for this to work it is necessary to mantain the
+        imported names unaltered. Since the text buffer is cleared on each
+        import only the last import will be exported correctly."""
+        # clear the text buffer, or create new buffer
+        try:
+            namestxt = bpy.data.texts["FullNames"]
+            namestxt.clear()
+        except KeyError:
+            namestxt = bpy.data.texts.new("FullNames")
+            
+        # write the names to the text buffer
+        for block, shortname in self.dict_names.items():
+            block_name = block.name.decode()
+            if block_name and shortname != block_name:
+                namestxt.write('%s;%s\n' % (shortname, block_name))
